Move pipeline tests to unit testing side
